+++ demo/07terrainheightmap/app.h
@@ -11,7 +11,7 @@ using std::string;
 bool load();
 void free();
 void init(Renderer &gfx, Context &ctx);
-void update(Renderer &gfx, Context &ctx, double dt);
-void render(Renderer &gfx, Context &ctx, double dt);
+void update(Renderer &gfx, Context &ctx, float dt);
+void render(Renderer &gfx, Context &ctx, float dt);
 
 #endif
